* Refactoring cloud-based solution. The approach I wanted to take didn't work
